feat: 숫자야구 클래스 분리
