+++ docs/README.md
@@ -1,0 +1,9 @@
+# 미션 - 숫자 야구
+
+## 구현할 기능
+
+- 맞춰야할 숫자를 랜덤하게 받아 저장하는 기능
+- 사용자가 숫자를 입력하는 기능
+- 입력한 숫자가 형식에 맞지 않으면 오류를 반환하는 기능
+- 사용자가 입력한 숫자와 비교하며 볼,스트라이크,낫싱을 판단하는 기능
+- 모든 숫자를 맞췄을 경우 게임을 종료하고 재시작 유무를 물어보고 다시 실행하거나 종료하는 기능
