+++ src/utils/CheckInput.js
@@ -1,7 +1,7 @@
 import { INPUT_LENGTH } from './Constants';
 
 export const checkIsReplay = (input) => {
-	if ([1, 2].includes(input))
+	if (!['1', '2'].includes(input))
 		throw new Error('[ERROR] 1 또는 2만 입력가능합니다.');
 };
 
@@ -10,15 +10,14 @@ export const checkIsPitch = (input) => {
 		throw new Error(`[ERROR] 입력가능한 길이는 ${INPUT_LENGTH}입니다.`);
 	}
 
-	const inputArray = input.split().forEach((element) => parseInt(element));
+	const inputArray = [];
+	input.split().forEach((element) => {
+		if (inputArray.includes(parseInt(element))) {
+			throw new Error('[ERROR] 중복된 숫자가 입력되었습니다.');
+		}
+		inputArray.push(parseInt(element));
+	});
 	if (inputArray.includes(NaN)) {
 		throw new Error('[ERROR] 숫자만 입력가능합니다.');
 	}
-	// if (inputArray!==[//]) {
-	// 	throw new Error('[ERROR] 중복된 숫자가 입력되었습니다.');
-	// }
-	// input.split('').forEach((element) => {
-	// 	const convertedNumber = parseInt(element);
-	// 	inputArray.push(convertedNumber);
-	// });
 };
